Move UI engine to DataStructureState base class
Delegated control panel, sliders, and input buffers to base class.

Added child class specific menu content and execution.

Cleaned up LinkedListState to use inherited UI logic and assets.

Fixed RNG bug for LinkedListState. Added head and tail pointers to LinkedListState

+++ src/UI.cpp
@@ -121,7 +121,7 @@ void button::Draw(Vector2 mousePos)
 
 bool button::isPressed(Vector2 mousePos, bool mousePressed)
 {
-    Rectangle rect = {position.x, position.y, static_cast<double>(texture.width), static_cast<double>(texture.height)};
+    Rectangle rect = { position.x, position.y, (float)texture.width, (float)texture.height };
     if (CheckCollisionPointRec(mousePos, rect) && mousePressed) return true;
     return false;
 }
